# Make tungstenite-rs more safe