This reduce number of usage of `unsafe` in crate to just 1.
Plus I made benchmark (you can find benchmark [here](https://github.com/Dushistov/tungstenite_apply_mask_bench) (**it requires nightly**) ),
and it should be also faster now:
| ptr offset  | fallback (ns) | old fast32 (ns) | new fast32 (ns) |
| ------------- | ------------- | ------------- |  ------------- |
| 0  | 24.727  | 11.963 | 7.9807 |
| 1  | 25.710  | 20.476 | 9.3833 |
| 2 | 16.160 | 20.569 | 9.1830 |
| 3 | 15.548 | 18.867 | 8.3566 |